+++ silo-core/contracts/Silo.sol
@@ -122,6 +122,11 @@ contract Silo is ISilo, ShareCollateralToken {
     }
 
     /// @inheritdoc ISilo
+    function getFractionsStorage() external view returns (Fractions memory fractions) {
+        fractions = Views.getFractionsStorage();
+    }
+
+    /// @inheritdoc ISilo
     function getCollateralAssets() external view virtual returns (uint256 totalCollateralAssets) {
         totalCollateralAssets = _totalAssets();
     }