+++ silo-core/test/foundry/SiloFactory/SiloFactoryValidateSiloInitDataTest.t.sol
@@ -9,7 +9,7 @@ import {VeSiloContracts} from "ve-silo/common/VeSiloContracts.sol";
 
 import {ISiloConfig} from "silo-core/contracts/interfaces/ISiloConfig.sol";
 import {ISiloFactory} from "silo-core/contracts/SiloFactory.sol";
-import {SiloFactoryDeploy} from "silo-core/deploy/SiloFactoryDeploy.s.sol";
+import {SiloFactoryVeSiloDeploy} from "silo-core/deploy/silo-factory/SiloFactoryVeSiloDeploy.s.sol";
 
 import {OracleMock} from "../_mocks/OracleMock.sol";
 
@@ -23,11 +23,11 @@ contract SiloFactoryValidateSiloInitDataTest is Test {
     address internal _feeDistributor = makeAddr("FeeDistributor");
 
     function setUp() public {
-        // Mock addresses that we need for the `SiloFactoryDeploy` script
+        // Mock addresses that we need for the `SiloFactoryVeSiloDeploy` script
         AddrLib.setAddress(VeSiloContracts.TIMELOCK_CONTROLLER, _timelock);
         AddrLib.setAddress(VeSiloContracts.FEE_DISTRIBUTOR, _feeDistributor);
 
-        SiloFactoryDeploy siloFactoryDeploy = new SiloFactoryDeploy();
+        SiloFactoryVeSiloDeploy siloFactoryDeploy = new SiloFactoryVeSiloDeploy();
         siloFactoryDeploy.disableDeploymentsSync();
         siloFactory = siloFactoryDeploy.run();
     }