+++ silo-core/test/foundry/utils/ShareTokenCommonTest.sol
@@ -326,15 +326,15 @@ contract ShareTokenCommonTest is SiloLittleHelper, Test, ERC20PermitUpgradeable 
     function test_shareTokenName() public view {
         (address protected0, address collateral0, address debt0) = siloConfig.getShareTokens(address(silo0));
 
-        assertEq(IShareToken(protected0).name(), "Silo Finance Non-borrowable b Deposit, SiloId: 1");
-        assertEq(IShareToken(collateral0).name(), "Silo Finance Borrowable b Deposit, SiloId: 1");
-        assertEq(IShareToken(debt0).name(), "Silo Finance b Debt, SiloId: 1");
+        assertEq(IShareToken(protected0).name(), "Silo Finance Non-borrowable b Deposit, SiloId: 100");
+        assertEq(IShareToken(collateral0).name(), "Silo Finance Borrowable b Deposit, SiloId: 100");
+        assertEq(IShareToken(debt0).name(), "Silo Finance b Debt, SiloId: 100");
 
         (address protected1, address collateral1, address debt1) = siloConfig.getShareTokens(address(silo1));
 
-        assertEq(IShareToken(protected1).name(), "Silo Finance Non-borrowable b Deposit, SiloId: 1");
-        assertEq(IShareToken(collateral1).name(), "Silo Finance Borrowable b Deposit, SiloId: 1");
-        assertEq(IShareToken(debt1).name(), "Silo Finance b Debt, SiloId: 1");
+        assertEq(IShareToken(protected1).name(), "Silo Finance Non-borrowable b Deposit, SiloId: 100");
+        assertEq(IShareToken(collateral1).name(), "Silo Finance Borrowable b Deposit, SiloId: 100");
+        assertEq(IShareToken(debt1).name(), "Silo Finance b Debt, SiloId: 100");
     }
 
     /*
@@ -343,15 +343,15 @@ contract ShareTokenCommonTest is SiloLittleHelper, Test, ERC20PermitUpgradeable 
     function test_shareTokenSymbol() public view {
         (address protected0, address collateral0, address debt0) = siloConfig.getShareTokens(address(silo0));
 
-        assertEq(IShareToken(protected0).symbol(), "nbb-1");
-        assertEq(IShareToken(collateral0).symbol(), "bb-1");
-        assertEq(IShareToken(debt0).symbol(), "db-1");
+        assertEq(IShareToken(protected0).symbol(), "nbb-100");
+        assertEq(IShareToken(collateral0).symbol(), "bb-100");
+        assertEq(IShareToken(debt0).symbol(), "db-100");
 
         (address protected1, address collateral1, address debt1) = siloConfig.getShareTokens(address(silo1));
 
-        assertEq(IShareToken(protected1).symbol(), "nbb-1");
-        assertEq(IShareToken(collateral1).symbol(), "bb-1");
-        assertEq(IShareToken(debt1).symbol(), "db-1");
+        assertEq(IShareToken(protected1).symbol(), "nbb-100");
+        assertEq(IShareToken(collateral1).symbol(), "bb-100");
+        assertEq(IShareToken(debt1).symbol(), "db-100");
     }
 
     /*