+++ silo-vaults/contracts/incentives/claiming-logics/SiloIncentivesControllerCLFactory.sol
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.28;
 
+import {Create2Factory} from "common/utils/Create2Factory.sol";
 import {ISiloIncentivesControllerCLFactory} from "../../interfaces/ISiloIncentivesControllerCLFactory.sol";
 import {SiloIncentivesControllerCL} from "./SiloIncentivesControllerCL.sol";
 
 /// @dev Factory for creating SiloIncentivesControllerCL instances
-contract SiloIncentivesControllerCLFactory is ISiloIncentivesControllerCLFactory {
+contract SiloIncentivesControllerCLFactory is Create2Factory, ISiloIncentivesControllerCLFactory {
     mapping(address => bool) public createdInFactory;
 
     /// @inheritdoc ISiloIncentivesControllerCLFactory
@@ -13,7 +14,7 @@ contract SiloIncentivesControllerCLFactory is ISiloIncentivesControllerCLFactory
         address _vaultIncentivesController,
         address _siloIncentivesController
     ) external returns (SiloIncentivesControllerCL logic) {
-        logic = new SiloIncentivesControllerCL(_vaultIncentivesController, _siloIncentivesController);
+        logic = new SiloIncentivesControllerCL{salt: _salt()}(_vaultIncentivesController, _siloIncentivesController);
 
         createdInFactory[address(logic)] = true;